+++ pkg/deployment/resources/pod_cleanup.go
@@ -17,8 +17,6 @@
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
-// Author Ewout Prangsma
-//
 
 package resources
 
@@ -26,14 +24,13 @@ import (
 	"context"
 	"time"
 
-	"github.com/arangodb/kube-arangodb/pkg/deployment/resources/inspector"
 	v1 "k8s.io/api/core/v1"
 
+	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
+	"github.com/arangodb/kube-arangodb/pkg/deployment/resources/inspector"
 	"github.com/arangodb/kube-arangodb/pkg/util"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
 	inspectorInterface "github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector"
-
-	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 )
 
 const (
@@ -55,15 +52,19 @@ func (r *Resources) CleanupTerminatedPods(ctx context.Context, cachedStatus insp
 			return nil
 		}
 
-		if !(k8sutil.IsPodSucceeded(pod) || k8sutil.IsPodFailed(pod) || k8sutil.IsPodTerminating(pod)) {
-			return nil
-		}
-
 		// Find member status
 		memberStatus, group, found := status.Members.MemberStatusByPodName(pod.GetName())
 		if !found {
 			log.Debug().Str("pod", pod.GetName()).Msg("no memberstatus found for pod. Performing cleanup")
 		} else {
+			spec := r.context.GetSpec()
+			coreContainers := spec.GetCoreContainers(group)
+			if !(k8sutil.IsPodSucceeded(pod, coreContainers) || k8sutil.IsPodFailed(pod, coreContainers) ||
+				k8sutil.IsPodTerminating(pod)) {
+				// The pod is not being terminated or failed or succeeded.
+				return nil
+			}
+
 			// Check member termination condition
 			if !memberStatus.Conditions.IsTrue(api.ConditionTypeTerminated) {
 				if !group.IsStateless() {