+++ pkg/deployment/resources/pod_creator.go
@@ -515,7 +515,6 @@ func (r *Resources) createPodForMember(ctx context.Context, cachedStatus inspect
 
 	imageInfo = *m.Image
 	apiObject := r.context.GetAPIObject()
-	groupSpec := spec.GetServerGroupSpec(group)
 
 	// Update pod name
 	role := group.AsRole()
@@ -543,11 +542,8 @@ func (r *Resources) createPodForMember(ctx context.Context, cachedStatus inspect
 		m.ArangoVersion = m.Image.ArangoDBVersion
 		m.ImageID = m.Image.ImageID
 
-		// Check for missing side cars in
-		m.SideCarSpecs = make(map[string]core.Container)
-		for _, specSidecar := range groupSpec.GetSidecars() {
-			m.SideCarSpecs[specSidecar.Name] = *specSidecar.DeepCopy()
-		}
+		// reset old sidecar values to nil
+		m.SideCarSpecs = nil
 
 		log.Debug().Str("pod-name", m.PodName).Msg("Created pod")
 		if m.Image == nil {